+++ client/src/components/atoms/CalendarDays/LayerDay/index.tsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const LayerDay = styled.div`
+  position: relative;
+  height: 8.8rem;
+  border-top: 1px solid #00000000;
+  overflow: hidden;
+`;
+
+export default LayerDay;
