+++ client/src/components/molecules/Calendars/LayerScheduleCalendar/index.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import LayerScheduleDay from './LayerScheduleDay';
+import { LayerScheduleDay } from 'components/atoms';
 import { useRecoilValue } from 'recoil';
 import { filteredTaskState } from 'recoil/project';
 import { IMindNode } from 'types/mindmap';
