+++ client/src/components/organisms/ScheduleCalendar/index.tsx
@@ -1,15 +1,11 @@
 import { useState } from 'react';
 import { CalendarWrapper, LayerWrapper, Wrapper } from './style';
-import MonthSelector from './MonthSelector';
-import Calendar from './Calendar';
-import LayerSprintMonth from './LayerSprintMonth';
-import LayerTaskDetailMonth from './LayerTaskDetailMonth';
-import LayerScheduleMonth from './LayerScheduleMonth';
 
 import { useSetRecoilState } from 'recoil';
 import { selectedNodeIdState } from 'recoil/node';
 import { IMindNode } from 'types/mindmap';
 import { getTodayISODate } from 'utils/date';
+import { Calendar, LayerScheduleCalendar, LayerSprintCalendar, LayerTaskDetailCalendar, MonthSelector } from 'components/molecules';
 
 const ScheduleCalendar = () => {
   const setSelectedNodeId = useSetRecoilState(selectedNodeIdState);
@@ -26,13 +22,13 @@ const ScheduleCalendar = () => {
         <Calendar currentDateISO={currentDateISO} />
       </CalendarWrapper>
       <LayerWrapper className={hoveredNode ? 'blur no-click' : 'no-click'}>
-        <LayerSprintMonth currentDateISO={currentDateISO} />
+        <LayerSprintCalendar currentDateISO={currentDateISO} />
       </LayerWrapper>
       <LayerWrapper className={hoveredNode ? 'blur' : ''}>
-        <LayerScheduleMonth currentDateISO={currentDateISO} setHoveredNode={setHoveredNode} />
+        <LayerScheduleCalendar currentDateISO={currentDateISO} setHoveredNode={setHoveredNode} />
       </LayerWrapper>
       <LayerWrapper className={'no-click'}>
-        <LayerTaskDetailMonth currentDateISO={currentDateISO} hoveredNode={hoveredNode} />
+        <LayerTaskDetailCalendar currentDateISO={currentDateISO} hoveredNode={hoveredNode} />
       </LayerWrapper>
     </Wrapper>
   );
