+++ client/src/components/organisms/ScheduleCalendar/style.ts
@@ -14,19 +14,6 @@ const Wrapper = styled.div`
   }
 `;
 
-const MonthSelectorWrapper = styled.div`
-  ${({ theme }) => theme.flex.center};
-  margin-top: 3rem;
-  color: ${({ theme }) => theme.color.primary2};
-  font-size: ${({ theme }) => theme.fontSize.title};
-  font-weight: bold;
-
-  span {
-    ${({ theme }) => theme.flex.center};
-    width: 20rem;
-  }
-`;
-
 const CalendarWrapper = styled.div`
   display: grid;
   grid-template-columns: repeat(7, 1fr);
@@ -37,34 +24,6 @@ const CalendarWrapper = styled.div`
   border-radius: 0.5rem;
   overflow: hidden;
   ${({ theme }) => theme.shadow};
-`;
-
-const CalendarHeader = styled.div`
-  ${({ theme }) => theme.flex.center};
-  height: 2rem;
-  color: ${({ theme }) => theme.color.white};
-  background-color: ${({ theme }) => theme.color.primary1};
-  border-right: 1px solid ${({ theme }) => theme.color.gray3};
-`;
-
-const CalendarDay = styled.div<{ disable?: boolean; today?: boolean }>`
-  position: relative;
-  height: 8.8rem;
-  padding: 0.2rem;
-  border-top: 1px solid ${({ theme }) => theme.color.gray3};
-  border-right: 1px solid ${({ theme }) => theme.color.gray3};
-  overflow: hidden;
-
-  color: ${({ today, theme }) => (today ? theme.color.primary1 : theme.color.gray1)};
-  background-color: ${({ disable, today, theme }) => {
-    if (disable) {
-      return theme.color.gray4;
-    }
-    if (today) {
-      return theme.color.primary3;
-    }
-    return '';
-  }};
 `;
 
 const LayerWrapper = styled.div`
@@ -82,121 +41,4 @@ const LayerWrapper = styled.div`
   overflow: hidden;
 `;
 
-const LayerDay = styled.div`
-  position: relative;
-  height: 8.8rem;
-  border-top: 1px solid #00000000;
-  overflow: hidden;
-`;
-
-const LayerScheduleDayWrapper = styled(LayerDay)`
-  height: 7.7rem;
-  margin-top: 1.1rem;
-  padding: 0.2rem;
-  overflow-y: auto;
-  ${({ theme }) => theme.customScrollbar.primary3}
-`;
-
-const LayerTask = styled.div<{ delayed?: boolean | null; ended?: boolean | null }>`
-  ${({ theme }) => theme.flex.rowCenter}
-
-  span {
-    font-weight: normal;
-    margin-bottom: 0.3rem;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-
-  cursor: pointer;
-
-  color: ${({ theme, delayed, ended }) => {
-    if (delayed) {
-      return theme.color.red;
-    }
-    if (ended) {
-      return theme.color.primary1;
-    }
-    return theme.color.gray1;
-  }};
-
-  :hover {
-    color: ${({ theme }) => theme.color.white};
-    background-color: ${({ theme }) => theme.color.primary1};
-  }
-`;
-
-const LayerTaskDetailDayWrapper = styled(LayerDay)`
-  padding-top: 3.8rem;
-  color: ${({ theme }) => theme.color.white};
-
-  div {
-    ${({ theme }) => theme.flex.center};
-    width: 100%;
-    height: 1.2rem;
-    border-radius: 0.5rem;
-  }
-
-  hr {
-    margin: 0.5rem 0.1rem;
-    border-top: dotted 2.5px;
-  }
-
-  .created {
-    background-color: ${({ theme }) => theme.color.gray1};
-  }
-  .started {
-    background-color: ${({ theme }) => theme.color.primary1};
-  }
-  .ended {
-    background-color: ${({ theme }) => theme.color.primary2};
-  }
-  .due {
-    background-color: ${({ theme }) => theme.color.red};
-  }
-  .waiting {
-    color: ${({ theme }) => theme.color.gray1};
-  }
-  .working {
-    color: ${({ theme }) => theme.color.primary1};
-  }
-  .resting {
-    color: ${({ theme }) => theme.color.primary2};
-  }
-  .delaying {
-    color: ${({ theme }) => theme.color.red};
-  }
-`;
-
-const LayerSprintDayWrapper = styled(LayerDay)`
-  opacity: 0.4;
-
-  .start {
-    border-top-left-radius: 0.5rem;
-    border-bottom-left-radius: 0.5rem;
-  }
-  .end {
-    border-top-right-radius: 0.5rem;
-    border-bottom-right-radius: 0.5rem;
-  }
-`;
-
-const LayerSprint = styled.div<{ color: string }>`
-  width: 100%;
-  height: 1.2rem;
-  background-color: ${({ color }) => `#${color}`}; ;
-`;
-
-export {
-  Wrapper,
-  MonthSelectorWrapper,
-  CalendarWrapper,
-  CalendarHeader,
-  CalendarDay,
-  LayerWrapper,
-  LayerScheduleDayWrapper,
-  LayerTask,
-  LayerTaskDetailDayWrapper,
-  LayerSprintDayWrapper,
-  LayerSprint,
-};
+export { Wrapper, CalendarWrapper, LayerWrapper };
